feat: turrets upgrade & buy system

+++ Weapon Tycoon/Assets/_Project/Scripts/Infrastructure/Data/Turrets/TurretUpgradeConfig.cs
@@ -13,7 +13,7 @@ namespace _Project.Scripts.Infrastructure.Data.Turrets
         public class TurretStat
         {
             public int Damage;
-            public float RPM;
+            public int RPM;
             public long BuyPrice;
         }
     }
